+++ src/input_output/FGOutputType.cpp
@@ -189,7 +189,7 @@ void FGOutputType::SetRateHz(double rtHz)
 {
   rtHz = rtHz>1000?1000:(rtHz<0?0:rtHz);
   if (rtHz > 0) {
-    SetRate(0.5 + 1.0/(FDMExec->GetDeltaT()*rtHz));
+    SetRate((unsigned int)(0.5 + 1.0/(FDMExec->GetDeltaT()*rtHz))); // cast away warning - RobD
     Enable();
   } else {
     SetRate(1);
